# docs/swagger.yaml
basePath: /
definitions:
...
  dto.DetailProduct:
    properties:
      description:
        type: string
      id_images:
        items:
          type: string
        type: array
      id_product:
        type: integer
      images:
        items:
          type: string
        type: array
      price:
        type: number
      product_name:
        type: string
    type: object
  dto.DetailProductUser:
    properties:
      description:
        type: string
      discount:
        type: number
      id_product:
        type: integer
      images:
        type: string
      price:
        type: number
      product_name:
        type: string
      rating:
        type: number
      total_review:
        type: integer
    type: object
...
  dto.PaginationMeta:
    properties:
      next_page:
        type: string
      page:
        type: integer
      prev_page:
        type: string
      total_page:
        type: integer
    type: object
  dto.ProductResponse:
    properties:
      data:
        items:
          $ref: "#/definitions/dto.Products"
        type: array
      message:
        example: Data retrieved successfully
        type: string
      meta:
        $ref: "#/definitions/dto.PaginationMeta"
      status:
        example: Success
        type: string
    type: object
  dto.ProductSize:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
    type: object
  dto.ProductType:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
    type: object
  dto.Products:
    properties:
      discount:
        type: number
      id:
        type: integer
      image_products:
        type: string
      name:
        type: string
      price:
        type: number
      rating_product:
        type: number
    type: object
...
  dto.ResponseError:
    properties:
      errors:
        example: failed get data
        type: string
      message:
        example: Failed get data
        type: string
      status:
        example: Error
        type: string
    type: object
  dto.ResponseSuccess:
    properties:
      message:
        example: Data retrieved successfully
        type: string
      status:
        example: Success
        type: string
    type: object
...
host: 192.168.50.221:8080
info:
  contact: {}
  description: Solid Coffee Backend RESTful API
  title: Solid Coffee Backend
  version: "1.0"
paths:
...
  /admin/products/:
    post:
      consumes:
        - multipart/form-data
      parameters:
        - collectionFormat: csv
          description: Product images
          in: formData
          items:
            type: file
          name: images_file
          required: true
          type: array
        - description: Products name
          in: formData
          name: product_name
          required: true
          type: string
        - description: Price
          in: formData
          name: price
          required: true
          type: number
        - description: Description
          in: formData
          name: description
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Post product
      tags:
        - Admin Product Management
  /admin/products/{id}:
    delete:
      consumes:
        - application/json
      parameters:
        - description: Product id
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Delete product
      tags:
        - Admin Product Management
    get:
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.DetailProduct"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Get detail products by id
      tags:
        - Admin Product Management
    patch:
      consumes:
        - multipart/form-data
      parameters:
        - description: Product Id
          in: path
          name: id
          required: true
          type: integer
        - collectionFormat: csv
          description: Product images
          in: formData
          items:
            type: file
          name: images_file
          type: array
        - description: Products name
          in: formData
          name: product_name
          type: string
        - description: Price
          in: formData
          name: price
          type: number
        - description: Description
          in: formData
          name: description
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Update product
      tags:
        - Admin Product Management
  /admin/products/image/{id}:
    delete:
      consumes:
        - application/json
      parameters:
        - description: Image id
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Delete product image
      tags:
        - Admin Product Management
...
  /products/:
    get:
      parameters:
        - description: Blacklist id product
          in: query
          name: id
          type: string
        - description: Page number
          in: query
          name: page
          type: string
        - description: Product title search
          in: query
          name: title
          type: string
        - description: Minimum price
          in: query
          name: min
          type: string
        - description: Maximum price
          in: query
          name: max
          type: string
        - collectionFormat: multi
          description: Categories filter
          in: query
          items:
            type: string
          name: category
          type: array
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ProductResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      summary: Get all products
      tags:
        - Products
  /products/{id}:
    get:
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.DetailProductUser"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      summary: Get detail products by id
      tags:
        - Products
  /products/product-sizes/:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/dto.ProductSize"
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      summary: Get all product sizes
      tags:
        - Products
  /products/product-types/:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/dto.ProductType"
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      summary: Get all product types
      tags:
        - Products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"